+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Fix exponential notation syntax. https://github.com/rescript-lang/rescript/pull/7174
 - Fix bug where a ref assignment is moved ouside a conditional. https://github.com/rescript-lang/rescript/pull/7176
 - Fix nullable to opt conversion. https://github.com/rescript-lang/rescript/pull/7193
+- Raise error when defining external React components with `@react.componentWithProps`. https://github.com/rescript-lang/rescript/pull/7217
 
 #### :house: Internal